Add feature where admins are able to view dining reviews, and reject or approve them by changing their status.

+++ src/main/java/com/example/Restaurant/Allergy/Review/Application/Repository/DiningReviewRepository.java
@@ -9,6 +9,6 @@ import java.util.Optional;
 
 public interface DiningReviewRepository extends CrudRepository<DiningReview, Integer> {
     List<DiningReview> findByReviewStatus(ReviewStatus reviewStatus);
-    List<DiningReview> findByRestaurantIdAndReviewStatus(Long restaurantId, ReviewStatus reviewStatus);
-    Optional<DiningReview> findByCreatorNameAndRestaurantId(String creatorName, Long restaurantId);
+    List<DiningReview> findByRestaurantIdAndReviewStatus(Integer restaurantId, ReviewStatus reviewStatus);
+    Optional<DiningReview> findByCreatorNameAndRestaurantId(String creatorName, Integer restaurantId);
 }
