+++ src/main/java/com/example/Restaurant/Allergy/Review/Application/Repository/RestaurantRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.Optional;
 
 
-public interface RestaurantRepository extends CrudRepository<Restaurant, Long> {
+public interface RestaurantRepository extends CrudRepository<Restaurant, Integer> {
     List<Restaurant> findByZipcode(Integer zipcode);
     Optional<Restaurant> findByZipcodeAndName(Integer zipcode, String name);
     List<Restaurant> findByZipcodeAndPeanutScoreNotNullOrderByPeanutScoreDesc(Integer zipcode);
